Solar System on Unity
This is an accurately scaled simulation of the Solar System made as a course project for a course on Classical Physics. The planets follow their true elliptical trajectories, and the sizes and distances have been taken relative to that of Earth. Each planet has two camera views, with a default view to visualize the entire simulation at once.